+++ Cohort/Infrastructure/Sweep/SoftDeleteSweepStrategy.cs
@@ -53,7 +53,7 @@ public sealed class SoftDeleteSweepStrategy : IRetentionSweepStrategy
 
     public Strategy HandlesStrategy => Strategy.SoftDelete;
 
-    public Task<int> PreviewAsync(
+    public Task<long> PreviewAsync(
         RetentionEntry entry,
         RetentionRule rule,
         RetentionResolutionContext ctx,
@@ -64,7 +64,7 @@ public sealed class SoftDeleteSweepStrategy : IRetentionSweepStrategy
         return core.PreviewAsync(entry, rule, ctx, conn, ct);
     }
 
-    public Task<int> CountHeldAsync(
+    public Task<long> CountHeldAsync(
         RetentionEntry entry,
         RetentionRule rule,
         RetentionResolutionContext ctx,
@@ -88,7 +88,7 @@ public sealed class SoftDeleteSweepStrategy : IRetentionSweepStrategy
         return core.SweepAsync(entry, rule, ctx, conn, transaction, ct, execution);
     }
 
-    public Task<int> PreviewEraseAsync(
+    public Task<long> PreviewEraseAsync(
         RetentionEntry entry,
         RetentionRule rule,
         ErasureSubjectPredicate predicate,
@@ -101,7 +101,7 @@ public sealed class SoftDeleteSweepStrategy : IRetentionSweepStrategy
         return core.PreviewEraseAsync(entry, rule, predicate, tenant, now, conn, ct);
     }
 
-    public Task<int> CountHeldForEraseAsync(
+    public Task<long> CountHeldForEraseAsync(
         RetentionEntry entry,
         RetentionRule rule,
         ErasureSubjectPredicate predicate,
